move coefficients tests to their own file
